+++ argopy/tests/test_fetchers_facade_index.py
@@ -11,15 +11,53 @@ from . import (
     requires_localftp_index,
     requires_connection,
     safe_to_server_errors,
-    ci_erddap_index
+    ci_erddap_index,
+    requires_matplotlib,
+    has_matplotlib,
+    has_seaborn,
+    has_cartopy
 )
+
+
+if has_matplotlib:
+    import matplotlib as mpl
+
+if has_cartopy:
+    import cartopy
+
 
 ERDDAP_TIMEOUT = 3 * 60
 
 
+@requires_connection
 class Test_Facade:
+    local_ftp = argopy.tutorial.open_dataset("localftp")[0]
+    src = 'localftp'
+    # src = list(AVAILABLE_INDEX_SOURCES.keys())[0]
 
-    src = list(AVAILABLE_INDEX_SOURCES.keys())[0]
+    def __get_fetcher(self, empty: bool = False, pt: str = 'profile'):
+        f = ArgoIndexFetcher(src=self.src)
+        # f.valid_access_points[0]
+
+        if pt == 'float':
+            if not empty:
+                return f, ArgoIndexFetcher(src=self.src).float(2901623)
+            else:
+                return f, ArgoIndexFetcher(src=self.src).float(12)
+
+        if pt == 'profile':
+            if not empty:
+                return f, ArgoIndexFetcher(src=self.src).profile(2901623, 12)
+            else:
+                return f, ArgoIndexFetcher(src=self.src).profile(12, 1200)
+
+        if pt == 'region':
+            if not empty:
+                return f, ArgoIndexFetcher(src=self.src).region([-60, -55, 40.0, 45.0, 0.0, 10.0,
+                                                                "2007-08-01", "2007-09-01"])
+            else:
+                return f, ArgoIndexFetcher(src=self.src).region([-60, -55, 40.0, 45.0, 99.92, 99.99,
+                                                                "2007-08-01", "2007-08-01"])
 
     def test_invalid_fetcher(self):
         with pytest.raises(InvalidFetcher):
@@ -41,6 +79,49 @@ class Test_Facade:
     def test_invalid_dataset(self):
         with pytest.raises(ValueError):
             ArgoIndexFetcher(src=self.src, ds='dummy_ds')
+
+    @requires_matplotlib
+    def test_plot(self):
+        with argopy.set_options(local_ftp=self.local_ftp):
+            f, fetcher = self.__get_fetcher(pt='float')
+
+            # Test 'trajectory'
+            for ws in [False, has_seaborn]:
+                for wc in [False, has_cartopy]:
+                    for legend in [True, False]:
+                        fig, ax = fetcher.plot(ptype='trajectory', with_seaborn=ws, with_cartopy=wc, add_legend=legend)
+                        assert isinstance(fig, mpl.figure.Figure)
+
+                        expected_ax_type = (
+                            cartopy.mpl.geoaxes.GeoAxesSubplot
+                            if has_cartopy and wc
+                            else mpl.axes.Axes
+                        )
+                        assert isinstance(ax, expected_ax_type)
+
+                        expected_lg_type = mpl.legend.Legend if legend else type(None)
+                        assert isinstance(ax.get_legend(), expected_lg_type)
+
+                        mpl.pyplot.close(fig)
+
+            # Test 'dac', 'profiler'
+            for ws in [False, has_seaborn]:
+                for by in [
+                    "dac",
+                    "profiler"
+                ]:
+                    fig, ax = fetcher.plot(ptype=by, with_seaborn=ws)
+                    assert isinstance(fig, mpl.figure.Figure)
+                    mpl.pyplot.close(fig)
+
+            # Test 'qc_altimetry'
+            import IPython
+            dsh = fetcher.plot(ptype='qc_altimetry', embed='slide')
+            assert isinstance(dsh(0), IPython.display.Image)
+
+            # Test invalid plot
+            with pytest.raises(ValueError):
+                fetcher.plot(ptype='invalid_cat', with_seaborn=ws)
 
 
 @requires_connection